requirements.txt
numpy
numpy-stl
scipy
matplotlib
trimesh